+++ lib/extensions/datetime_ext.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: non_constant_identifier_names
 
 import 'package:intl/intl.dart';
+import 'package:s_extensions/extensions/number_ext.dart';
 
 extension DateTimeExtensions on DateTime {
   // ? =================== Formaters ========================
@@ -14,26 +15,62 @@ extension DateTimeExtensions on DateTime {
   /// Returns a string representation of this date with the format
   /// `yyyy-MM-dd`.
   ///
-  /// Example: `DateTime.now().yyyyMMdd` -> `2023-08-24`.
-  String get yyyyMMdd => format('yyyy-MM-dd');
+  /// Example: `DateTime.now().formatYearMonthDay` -> `2023-08-24`.
+  String get formatYearMonthDay => format('yyyy-MM-dd');
+
+  /// Returns a string representation of this date with the format
+  /// `dd-MM-yyyy`.
+  ///
+  /// Example: `DateTime.now().formatDayMonthYear` -> `24-08-2023`.
+  String get formatDayMonthYear => format('dd-MM-yyyy');
 
   /// Returns a string representation of this date with the format
   /// `yyyy-MM-dd HH:mm:ss`.
   ///
-  /// Example: `DateTime.now().yyyyMMddHHmmss` -> `2023-08-24 12:34:56`.
-  String get yyyyMMddHHmmss => format('yyyy-MM-dd HH:mm:ss');
+  /// Example: `DateTime.now().formatYearMonthDayHrMinSec` -> `2023-08-24 12:34:56`.
+  String get formatYearMonthDayHrMinSec => format('yyyy-MM-dd hh:mm:ss');
 
   /// Returns a string representation of this date with the format
   /// `yyyy-MM-dd HH:mm`.
   ///
-  /// Example: `DateTime.now().yyyyMMddHHmm` -> `2023-08-24 12:34`.
-  String get yyyyMMddHHmm => format('yyyy-MM-dd HH:mm');
+  /// Example: `DateTime.now().formatYearMonthDayHrMin` -> `2023-08-24 12:34`.
+  String get formatYearMonthDayHrMin => format('yyyy-MM-dd hh:mm');
 
   /// Returns a string representation of this date with the format
-  /// `HH:mm`.
+  /// `HH:mm` in 12 hour format.
   ///
-  /// Example: `DateTime.now().HHmm` -> `12:34`.
-  String get HHmm => format('HH:mm');
+  /// Example: `DateTime.now().formatHrMin12` -> `12:34 PM`.
+  String get formatHrMin12 => format('HH:mm a');
+
+  /// Returns a string representation of this date with the format
+  /// `HH:mm` in 24 hour format.
+  ///
+  /// Example: `DateTime.now().formatHrMin24` -> `15:34`.
+  String get formatHrMin24 => format('hh:mm');
+
+  /// Returns a string representation of this date with the format
+  /// `HH:mm` in 12 hour format.
+  ///
+  /// Example: `DateTime.now().formatHrMinSec12` -> `12:34:56 PM`.
+  String get formatHrMinSec12 => format('HH:mm:ss a');
+
+  /// Returns a string representation of this date with the format
+  /// `HH:mm` in 24 hour format.
+  ///
+  /// Example: `DateTime.now().formatHrMinSec24` -> `15:34:56`.
+  String get formatHrMinSec24 => format('hh:mm:ss');
+
+  /// Returns a string representation of this date with the format
+  /// `dd`.
+  ///
+  /// Example: `DateTime.now().formatDay` -> `Wednesday`.
+  String get formatDay => format('EEEE');
+
+  /// Returns a string representation of this date with the format
+  /// `MMMM`.
+  ///
+  /// Example: `DateTime.now().formatMonth` -> `August`.
+  String get formatMonth => format('MMMM');
 
   // ? =================== Validators ========================
 
@@ -50,19 +87,19 @@ extension DateTimeExtensions on DateTime {
   /// Returns `true` if this date is today.
   ///
   /// Example: `DateTime.now().isToday` -> `true`.
-  bool get isToday => yyyyMMdd == DateTime.now().yyyyMMdd;
+  bool get isToday => formatYearMonthDay == DateTime.now().formatYearMonthDay;
 
   /// Returns `true` if this date is tomorrow.
   ///
   /// Example: `DateTime.now().isTomorrow` -> `true`.
   bool get isTomorrow =>
-      yyyyMMdd == DateTime.now().add(const Duration(days: 1)).yyyyMMdd;
+      formatYearMonthDay == (DateTime.now() + 1.day).formatYearMonthDay;
 
   /// Returns `true` if this date is yesterday.
   ///
   /// Example: `DateTime.now().isYesterday` -> `true`.
   bool get isYesterday =>
-      yyyyMMdd == DateTime.now().subtract(const Duration(days: 1)).yyyyMMdd;
+      formatYearMonthDay == (DateTime.now() - 1.day).formatYearMonthDay;
 
   /// Returns `true` if this date is in the previous month.
   ///
@@ -92,12 +129,12 @@ extension DateTimeExtensions on DateTime {
 
   /// Ads this duration of time to this date and returns the result.
   ///
-  /// Example: `"2023-08-24".toDate + 7.day` -> `2023-08-31 00:00:00`.
+  /// Example: `"2023-08-24".toDate + 7.day` -> `2023-08-31`.
   DateTime operator +(Duration duration) => add(duration);
 
   /// Subtracts this duration of time from this date and returns the result.
   ///
-  /// Example: `"2023-08-24".toDate - 7.day` -> `2023-08-17 00:00:00`.
+  /// Example: `"2023-08-24".toDate - 7.day` -> `2023-08-17`.
   DateTime operator -(Duration duration) => subtract(duration);
 
   /// Returns the duration between this date and now.